refactor: rocksdb separate file
