+++ components/ordhook-core/src/service/mod.rs
@@ -22,13 +22,16 @@ use crate::core::protocol::inscription_parsing::{
 };
 use crate::core::protocol::inscription_sequencing::SequenceCursor;
 use crate::core::{new_traversals_lazy_cache, should_sync_ordhook_db, should_sync_rocks_db};
+use crate::db::blocks::{
+    find_missing_blocks, insert_entry_in_blocks, open_ordhook_db_conn_rocks_db_loop, run_compaction,
+};
+use crate::db::update_sequence_metadata_with_block;
 use crate::db::{
     delete_data_in_ordhook_db, find_latest_inscription_block_height,
-    get_latest_indexed_inscription_number, insert_entry_in_blocks,
-    open_ordhook_db_conn_rocks_db_loop, open_readonly_ordhook_db_conn, open_readwrite_ordhook_dbs,
-    update_ordinals_db_with_block, BlockBytesCursor, TransactionBytesCursor,
+    get_latest_indexed_inscription_number, open_readonly_ordhook_db_conn,
+    open_readwrite_ordhook_dbs, update_ordinals_db_with_block, BlockBytesCursor,
+    TransactionBytesCursor,
 };
-use crate::db::{find_missing_blocks, run_compaction, update_sequence_metadata_with_block};
 use crate::scan::bitcoin::process_block_with_predicates;
 use crate::service::observers::create_and_consolidate_chainhook_config_with_predicates;
 use crate::service::runloops::start_bitcoin_scan_runloop;
